+++ apps/gnome/themeLoader.js
@@ -206,7 +206,6 @@ function localizeBuiltInTheme(theme, i18n) {
         return theme;
     const key = {
         'fallout-2': 'fallout2',
-        'fallout-3': 'fallout3',
         'night-video-deck': 'nightVideoDeck',
     }[theme.id];
     if (!key)